Add a pluggable callback API for custom signers.
This adds an API for caller code to provide signing functions, using the Go
core library http.HandleFunc idiom. As part of this, the former `main()` in
`warden.go` was refactored into a package intended to be used by an actual
`main()` implementation that sets up the binary with the desired set of
signing functions.

A new `default.go` is provided to illustrate usage.