Integrate APK signing into web server.
Adds a playground/warden/signfuncs handler for APK signing, and concomitant
sample config.

While we're at it, also adds a set of test certificates so that the demo
config works out of the box.

Also a couple tweaks to certificate comparison: switches to just comparing the
SHAs of the certs, since inspecting subject identity info is squirrelly, and
we only care about literally identical certs, anyway.