Support Android OTA signing.
This adds signing endpoints for Android A/B OTA images. Specifically, this
adds endpoints that know how to sign Android verified-boot images, dm-verity
(non-boot) partitions, and Brillo A/B-system payloads.

Also renames the vendored copy of the actual signing code, as it was
refactored a bit under a common 'playground/android' namespace.